### 2024-05-07 #6727 @editorialbot
[PRE REVIEW]: GRBoondi: A code for evolving Generalized Proca theories on arbitrary backgrounds
**Submitting author:** @ShaunFell (<a href="http://orcid.org/0000-0002-8059-0359">Shaun D. B. Fell</a>)
**Repository:** https://github.com/ShaunFell/GRBoondi
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
