+++ paper/paper.md
@@ -1,0 +1,91 @@
+---
+title: 'GRBoondi: A code for evolving Generalized Proca theories on arbitrary backgrounds'
+tags:
+  - c++
+  - MPI
+  - Open MP
+  - vector intrinsics
+  - Generalized Proca
+  - gravity
+  - general relativity
+  - numerical relativity
+  - adaptive refinement
+authors:
+- name: Shaun David Brocus Fell
+  orcid: 0000-0002-8059-0359
+  affiliation: 1
+- name: Lavinia Heisenberg
+  affiliation: 1
+affiliations:
+- name: Institute for Theoretical Physics, Universitaet Heidelberg, Philosophenweg 12, 69120 Heidelberg, Germany
+  index: 1
+date: 03-05-2024
+bibliography: paper.bib
+---
+
+# Summary
+
+Generalized Proca theories [@Heisenberg_2019;@Heisenberg_2014] provide a rich landscape to search for solutions to deep, fundamental questions, such as the nature of dark energy and dark matter. Moreover, strong gravity regimes, like those surrounding dense, compact astrophysical entities, offer novel avenues to explore fundamental fields. To effectively probe these domains, precise models are imperative for sifting through vast quantities of data. Crafting concrete models for the entirety of Generalized Proca theories represents a formidable endeavor, typically relying on numerical methods. These numerical methods are usually written to be case dependent, especially for a particular background. This makes it difficult to generalize the computational code to account for higher-order couplings or different types of backgrounds. The difficulty is especially amplified when considering the full landscape of Generalized Proca theories and beyond [@Heisenberg:2016eld]. 
+
+GRBoondi provides a unified interface for computing the evolution of any Generalized Proca model on an arbitrary background. Given a specific background with known expressions for the metric variables and initial data for the Proca field, GRBoondi numerically computes the time evolution of the Proca field, given user-specified Generalized Proca equations of motion. While the simultaneous evolution of the metric and matter fields offers the most comprehensive depiction of their temporal progression, there are instances where the density of the Proca field is negligible in comparison to the background curvature. In such cases, fixed background evolution routines serve as an excellent approximation to the complete mutual evolution. A core feature of GRBoondi is that it simplifies many of the boiler-plate code required to begin a simulation from scratch. The only additions a user needs to input are the initial conditions, the modifications to the equations of motion (EOM), and the background functions. GRBoondi will automatically compute various diagnostic quantities and plot files. Moreover, GRBoondi is incredibly modular and modifications to any of the in-built functions is effortless.
+
+The fixed background approximation provides a huge speedup compared to the full evolution, since only a small fraction of the variables actually require a time integration. Not only do simulations now require far less resources, the speed up can be up to hundreds of times faster. GRBoondi is based on the publicly available numerical relativity (NR) code GRChombo [@Andrade2021;@Clough2015sqa], which itself is based on the open source adaptive mesh refinement (AMR)-based differential equation solver Chombo [@Adams:2015kgr]. Some minor pieces of GRBoondi are also based off the recently released open source software GRDzhadzha [@Aurrekoetxea2024], which is also based off of GRChombo.
+
+# Statement of need
+
[…]
