+++ paper/paper.bib
@@ -1,0 +1,591 @@
+@article{Heisenberg_2019,
+	title        = {A systematic approach to generalisations of General Relativity and their cosmological implications},
+	author       = {Heisenberg, Lavinia},
+	year         = 2019,
+	month        = mar,
+	journal      = {Physics Reports},
+	volume       = 796,
+	pages        = {1–113},
+	doi          = {10.1016/j.physrep.2018.11.006},
+	issn         = {03701573},
+	note         = {arXiv: 1807.01725},
+	abstractnote = {A century ago, Einstein formulated his elegant and elaborate theory of General Relativity, which has so far withstood a multitude of empirical tests with remarkable success. Notwithstanding the triumphs of Einstein’s theory, the tenacious challenges of modern cosmology and of particle physics have motivated the exploration of further generalised theories of spacetime. Even though Einstein’s interpretation of gravity in terms of the curvature of spacetime is commonly adopted, the assignment of geometrical concepts to gravity is ambiguous because General Relativity allows three entirely diﬀerent, but equivalent approaches of which Einstein’s interpretation is only one. From a ﬁeld-theoretical perspective, however, the construction of a consistent theory for a Lorentz-invariant massless spin-2 particle uniquely leads to General Relativity. Keeping Lorentz invariance then implies that any modiﬁcation of General Relativity will inevitably introduce additional propagating degrees of freedom into the gravity sector. Adopting this perspective, we will review the recent progress in constructing consistent ﬁeld theories of gravity based on additional scalar, vector and tensor ﬁelds. Within this conceptual framework, we will discuss theories with Galileons, with Lagrange densities as constructed by Horndeski and beyond, extended to DHOST interactions, or containing generalized Proca ﬁelds and extensions thereof, or several Proca ﬁelds, as well as bigravity theories and scalar-vector-tensor theories. We will review the motivation of their inception, diﬀerent formulations, and essential results obtained within these classes of theories together with their empirical viability.},
+	language     = {en}
+}
+@article{Heisenberg_2014,
+	title        = {Generalization of the Proca Action},
+	author       = {Heisenberg, Lavinia},
+	year         = 2014,
+	month        = may,
+	journal      = {Journal of Cosmology and Astroparticle Physics},
+	volume       = 2014,
+	number       = {05},
+	pages        = {015–015},
+	doi          = {10.1088/1475-7516/2014/05/015},
+	issn         = {1475-7516},
+	note         = {arXiv: 1402.7026},
[…]
