### 2024-05-08 89de281
Fix DOIs in paper bib
