+++ paper/paper.bib
@@ -124,7 +124,7 @@
 	journal      = {Physics of the Dark Universe},
 	volume       = 32,
 	pages        = 100819,
-	doi          = {https://doi.org/10.1016/j.dark.2021.100819},
+	doi          = {10.1016/j.dark.2021.100819},
 	issn         = {2212-6864},
 	url          = {https://www.sciencedirect.com/science/article/pii/S2212686421000509},
 	abstract     = {We consider a specific dark energy model, which only includes the Lagrangian up to the cubic order in terms of the vector field self-interactions in the generalized Proca theory. We examine the cosmological parameters in the model by using the data sets of CMB and CMB+HST, respectively. In particular, the Hubble constant is found to be H0=71.80−0.72+1.07 (72.48−0.60+0.72) kms−1Mpc−1 at 68% C.L. with CMB (CMB+HST), which would alleviate the Hubble constant tension. We also obtain that the reduced χ2 values in our model are close to unity when fitting with CMB and CMB+HST, illustrating that our model is a good candidate to describe the cosmological evolutions of the universe.}
@@ -184,7 +184,8 @@
 	publisher    = {Springer},
 	address      = {New York, NY},
 	series       = {Lecture notes in physics},
-    journal      = ""
+	doi            = {10.1007/978-3-642-24525-1},
+    journal      = "Springer"
 }
 @article{Clough2015sqa,
 	title        = {{GRChombo : Numerical Relativity with Adaptive Mesh Refinement}},
@@ -448,7 +449,8 @@
 	year         = 2022,
 	eprint       = {2205.07784},
 	archiveprefix = {arXiv},
-	primaryclass = {gr-qc}
+	primaryclass = {gr-qc},
+	doi 				= {10.48550/arXiv.2205.07784}
 }
 @article{Hunter:2007,
 	title        = {Matplotlib: A 2D graphics environment},
