### 2024-05-08 2ab90f0
added DOI's to paper bib
