+++ paper/paper.bib
@@ -35,6 +35,7 @@
 	publisher    = {arXiv},
 	number       = {arXiv:2206.14320},
 	url          = {http://arxiv.org/abs/2206.14320},
+	doi 		= {10.48550/arXiv.2206.14320},
 	language     = {en},
     journal      = "arXiv"
 }
@@ -202,12 +203,12 @@
 	reportnumber = {KCL-PH-TH-2015-40}
 }
 @article{Adams:2015kgr,
-	title        = {{Chombo software package for AMR applications - design document}},
-	author       = {Adams, M. and others},
-	year         = 2015,
-	month        = 12,
-	reportnumber = {LBNL-6616E},
-    journal      = ""
+    author = "Adams, M. and Colella, P. and Graves, D.~T. and  Johnson, J.~N and Keen, N.~D. and Ligocki, T.~J. and Martin, D.~F. and McCorquodale, P.~W. and  Modiano, D. and Schwartz, P.~O and Sternberg T.~D. and Van Straalen, B. ",
+    title = "{Chombo software package for AMR applications - design document}",
+    reportNumber = "LBNL-6616E",
+    month = "12",
+    year = "2015",
+    url= "https://commons.lbl.gov/display/chombo"
 }
 @article{Aurrekoetxea2024,
 	title        = {GRDzhadzha: A code for evolving relativistic matter on analytic metric backgrounds},
@@ -240,14 +241,20 @@
 	publisher    = {Springer},
 	address      = {Berlin},
 	url          = {http://edoc.mpg.de/3341},
+	doi 		= {10.1007/3-540-36569-9_13},
 	suggested-for = {Cactus/Cactus}
 }
-@misc{Kranc:web,
-	title        = {{Kranc}: {Kranc} Assembles Numerical Code},
-	url          = {http://kranccode.org/},
-	suggested-for = {McLachlan/ML_BSSN},
-	key          = {Kranc},
-    year         = ""
+@article{Kranc:web,
+    author = "Husa, Sascha and Hinder, Ian and Lechner, Christiane",
+    title = "{Kranc: A Mathematica application to generate numerical codes for tensorial evolution equations}",
+    eprint = "gr-qc/0404023",
+    archivePrefix = "arXiv",
+    reportNumber = "AEI-2004-031",
+    doi = "10.1016/j.cpc.2006.02.002",
+    journal = "Comput. Phys. Commun.",
+    volume = "174",
+    pages = "983--1004",
+    year = "2006"
 }
 @article{Sperhake:2006cy,
 	title        = {{Binary black-hole evolutions of excision and puncture data}},
