### 2024-05-08 #6727 @dfm
@ShaunFell — Thank you for your submission! It seems like a large fraction of the code included in this repo is a copy of GRChombo which was [previously published in JOSS](https://github.com/openjournals/joss-reviews/issues/3703). Could you clarify which parts of the project are new to this submission?

In the meantime, all suitable editors are working at full capacity so I will "waitlist" this submission until someone is available. Thanks for your patience!
