### 2024-05-09 #6727 @ShaunFell
For future reviewers, to gauge the size of GRBoondi itself without the GRChombo source code, I run
`cloc --exclude-dir=GRChombo GRBoondi`
which produces:

`
     202 text files.
     161 unique files.
      78 files ignored.

Language                     files          blank        comment           code
C/C++ Header                    95           1747           1762           7258
C++                              9            349            261           2030
Python                           3            181            140            621
YAML                             5             51              7            316
make                            10             83             51            198
INI                              3             53             65            165
Markdown                         1             11              0             16
SUM:                           126           2475           2286          10604

`
