### 2024-05-23 359b578
add CITATION.cff file
