### 2024-05-25 0651873
replaced nested std::vector objects with Tensor objects (#9)
* replaced nested std::vector objects with Tensor objects

* clang format
