### 2024-05-25 a588346
removed double code line in KerrdeSitter
