### 2024-05-27 26a60d6
Forgot to remove debugging output in ExcisionEvolution
