+++ Source/Core/runGRChombo.hpp
@@ -64,8 +64,7 @@ template <class level_t> int runGRChombo(int argc, char *argv[])
     if (sim_params.run_initial_posttimestep)
     { // call the PostTimeStep right now!!!!
 
-        auto task = [](GRAMRLevel *level)
-        {
+        auto task = [](GRAMRLevel *level) {
             if (level->time() == 0.)
             {
                 level->specificPostTimeStep();
