### 2024-05-27 89f9c3e
Forgot to remove debugging output in ExcisionEvolution (#11)
* Forgot to remove debugging output in ExcisionEvolution

* clang format

* clang format with version 13.0.0
