+++ Source/GRChombo/Source/GRChomboCore/BoundaryConditions.cpp
@@ -572,8 +572,8 @@ void BoundaryConditions::fill_boundary_cells_dir(
                 MayDay::Error(
                     "BoundaryCondition::Supplied boundary not supported.");
             } // end switch
-        } // end iterate over box
-    } // end iterate over boxes
+        }     // end iterate over box
+    }         // end iterate over boxes
 }
 
 void BoundaryConditions::fill_sommerfeld_cell(
@@ -819,10 +819,10 @@ void BoundaryConditions::copy_boundary_cells(const Side::LoHiSide a_side,
                             m_dest_box(iv, icomp) = a_src[dind](iv, icomp);
                         }
                     } // end iterate over box
-                } // end iterate over boxes
-            } // end if(not periodic)
-        } // end iterate over spacedims
-    } // end test for same box layout
+                }     // end iterate over boxes
+            }         // end if(not periodic)
+        }             // end iterate over spacedims
+    }                 // end test for same box layout
 }
 
 /// Fill the fine boundary values in a_state
@@ -957,9 +957,9 @@ void BoundaryConditions::interp_boundaries(GRLevelData &a_fine_state,
                         local_stencil.fillFine(m_fine_box, m_coarse_box, iv);
                     }
                 } // end loop box
-            } // end loop boxes
-        } // end if is_periodic
-    } // end loop idir
+            }     // end loop boxes
+        }         // end if is_periodic
+    }             // end loop idir
 }
 
 /// Get the boundary condition for a_dir and a_side
