### 2024-05-27 cec2fa5
removed requirement for parameter file in SimpleDataReaderTest
