+++ Tests/SimpleDataReader_test/Main_datareader.cpp
@@ -48,11 +48,7 @@ static bool compareVectors(const vector<T> &a, const vector<T> &b)
 
 int main(int argc, char *argv[])
 {
-    char *dummy_argv[] = {
-	(char*)argv[0],
-	(char*)"2",
-	NULL
-    };
+    char *dummy_argv[] = {(char *)argv[0], (char *)"2", NULL};
     mainSetup((int)2, dummy_argv); // setup MPI processes
 
     int reader_failed{0};
