+++ .github/workflows/clang-format.yml
@@ -11,7 +11,7 @@ jobs:
 
     steps:
     - uses: actions/checkout@v4
-    - uses: DoozyX/clang-format-lint-action@v0.16.2
+    - uses: DoozyX/clang-format-lint-action@v0.17
       with:
         source: '.'
         extensions: 'hpp,cpp'
