### 2024-06-16 #6888 @Sbozzolo
Hi @ShaunFell,

I started my review and everything is looking great.

Could you please clearify the relationship between GRBoondi and GRChombo?

In terms of functionalities provided, my first impression is that GRBoondi would be optimally suited to be a module for GRChombo instead of being a separate codebase (GRBoondi even comes with a complete copy of GRChombo). Is there any intention to make GRBoondi an official module for GRChombo? Or, should GRBoondi be considered a hard fork of GRChombo? Having GRBoondi as a module would help with   shared tooling (e.g. post-processing) and reduce the maintenance burden on your end.

It would be really useful to have an end-to-end tutorial that explains how to run a simple example GRBoondi. Right now, most of the "getting started" is spread across multiple wiki pages and it is a little intimidating to figure out how to run a simple case.

I'd suggest adding a page with a tutorial to run a simple example. In this tutorial, you can focus on just running something and making a simple visualization. Then, you can provide backlinks to the pages on "Parameter files", "Running Examples", "Visualization" for more in depth discussions.
