### 2024-06-16 ShaunFell/GRBoondi#14 @Sbozzolo
Documentation improvements
Here are suggestions to improve the documentation:
- [x] One of the selling features for GRBoondi is that only a handful of files are required to simulate new interesting Proca scenarios. However, how to set these new cases up is not explained. Maybe you could add a tutorial on how to add a new case? This is different from an example because it guides the user in the process and highlights the various components (e.g., how does one fill the `params` folder? What is the role of `GNUmakefile`?), as opposed to just having all the files in one place
- [x] See comment in https://github.com/openjournals/joss-reviews/issues/6888#issuecomment-2171898643
- [x] Not related to documentation, but why is `csh` required? It is a little odd requirement
- [x] I think it would be useful to document the typical computational requirements for a simulation. E.g., how much RAM, disk, CPUs for a reasonable simulation
- [x] Documentation on parameters is a little overwhelming. If some parameters are more important than others, I'd suggest highlighting those to increase the signal-to-noise ratio
- [x] The Home page of the documentation does not take me to the next step. It should be clear where to click next.
- [x] There are no contributor guidelines, e.g., code style, contribution workflow, testing requirements
- [x] There is no documentation on how to run tests
- [x] There are not many backlinks to GRChombo and Chombo.
- [x] If GRBoondi is not supposed to be used as a black box, how are user supposed to learn how the internals work? There is no documentation about this
