+++ Examples/EMMinkowski/Main_EMKerrBH.cpp
@@ -3,22 +3,22 @@ Main method to run simulation
 */
 #include "EMFieldLevel.hpp"   // new Level class
 #include "SetupFunctions.hpp" //For setting up MPI processes
-#include "runGRChombo.hpp"    //main run function
+#include "runGRBoondi.hpp"    //main run function
 
 int main(int argc, char *argv[])
 {
     mainSetup(argc, argv); // setup MPI processes
 
-    int status = runGRChombo<EMFieldLevel>(
+    int status = runGRBoondi<EMFieldLevel>(
         argc, argv); // run simulation with modified level class
 
     if (status == 0)
     {
-        pout() << "GRChombo finished." << std::endl;
+        pout() << "GRBoondi finished." << std::endl;
     }
     else
     {
-        pout() << "GRChombo failed with return code " << status << std::endl;
+        pout() << "GRBoondi failed with return code " << status << std::endl;
     }
 
     mainFinalize(); // cleanup MPI processes
