### 2024-06-24 #6888 @KAClough
Hi @ShaunFell I haven't had a chance to go through the code but I wanted to echo what @Sbozzolo said about the relationship between GRChombo and GRBoondi not being very clear.

I would strongly suggest that you don't duplicate code from the GRChombo or GRDzhadzha repositories - it is fine for this code to have dependencies on them, in the same way as GRDzhadzha does on GRChombo, and this allows updates and optimisations that we make there to flow into your code. If there are aspects you found you needed to change in those repositories, I'd be happy to discuss making changes to them there to ensure compatibility, or finding a workaround.

If you agree I would suggest that this is amended before we review the code in more detail.
