### 2024-06-24 #6888 @ShaunFell
Hi @KAClough!
Ok, yea I originally included the GRChombo source because I needed to modify the level factory class, but I ended up implementing my own, so removing GRChombo in the source and instead using environment variables to point to the source isnt really an issue for me. I will remove the source from the repo and update the main branch accordingly. Thank you for the suggestion!
