### 2024-06-24 #6888 @ShaunFell
Hi @Sbozzolo!
Thank you for your feedback. I will go through the issues in the repo and respond accordingly in the next week or so. Just for transparency, I am writing my thesis, so my responses will be longer than usual.

> Could you please clearify the relationship between GRBoondi and GRChombo?

GRBoondi utilizes a lot of the source code from GRChombo, but also defines it's own functions for specifying the physical problems to that of generalized Proca around arbitrary spacetimes. While GRBoondi utilizes a lot of the GRChombo source, like the AMR interpolator and level classes, it also defines many of its own functionality, such as BaseProcaFieldLevel, since generalized Proca problems typically all share very similar structures when it comes to the code, like computing surface and volume integrals of the Proca norm, the Eulerian energy, etc. So broadly speaking, GRBoondi is very much it's own program that uses the AMR, level class, boundary code, etc. functionality of GRChombo/Chombo.

> In terms of functionalities provided, my first impression is that GRBoondi would be optimally suited to be a module for GRChombo instead of being a separate codebase (GRBoondi even comes with a complete copy of GRChombo). Is there any intention to make GRBoondi an official module for GRChombo? Or, should GRBoondi be considered a hard fork of GRChombo? Having GRBoondi as a module would help with shared tooling (e.g. post-processing) and reduce the maintenance burden on your end.

Sort of extending on what @KAClough mentioned, the GRChombo source will be removed from the GRBoondi source code. I did this since I had to modify some of the source code, but ended up creating my own classes and just left GRChombo in the GRBoondi source. I will remove this shortly with a pull request. So GRBoondi will required a fork of GRChombo. At the moment, I have no intention of making GRBoondi a module of GRChombo, though thats an interesting idea! generalized Proca theories are quite complex so keeping GRBoondi separate from GRChombo allows me to be quite versatile in changing the code for the needs of higher-order theories in the generalized Proca landscape.

> It would be really useful to have an end-to-end tutorial that explains how to run a simple example GRBoondi. Right now, most of the "getting started" is spread across multiple wiki pages and it is a little intimidating to figure out how to run a simple case.

ok, great idea! I will implement such a tutorial then.

> I'd suggest adding a page with a tutorial to run a simple example. In this tutorial, you can focus on just running something and making a simple visualization. Then, you can provide backlinks to the pages on "Parameter files", "Running Examples", "Visualization" for more in depth discussions.

Great idea, thanks!

> Could you also elaborate on the differences with `GRDzhadzha`?

GRDzhadzha is quite a generic toolkit for simulating any sort of matter field on generic fixed spacetimes. If one wants to simulate generalized Proca, they run into the same sort of problem as if they were to use GRChombo. Besides the lack of post-processing tools, users would have to manually rewrite a lot of the boilerplate code for each generalized Proca theory they simulate. This is basically the whole reason GRBoondi was created, to simplify a lot of this boilerplate code, making it much easier for users to study the generalized Proca landscape. For example, BaseProcaFieldLevel contains almost all the level class code that is common to all generalized Proca theories. This motif is shared throughout many of the classes in GRBoondi.
