### 2024-06-24 3e1e11d
Removed GRChombo source from source directory and added details in makefiles to set GRCHOMBO_SOURCE variable
