+++ Source/GRChombo/Source/GRChomboCore/ChomboParameters.hpp
@@ -1,607 +1,0 @@
-/* GRChombo
- * Copyright 2012 The GRChombo collaboration.
- * Please refer to LICENSE in GRChombo's root directory.
- */
-
-#ifndef CHOMBOPARAMETERS_HPP_
-#define CHOMBOPARAMETERS_HPP_
-
-// Chombo includes
-#include "MayDay.H"
-#include "Misc.H"
-
-// General includes
-#include "ArrayTools.hpp"
-#include "BoundaryConditions.hpp"
-#include "FilesystemTools.hpp"
-#include "GRParmParse.hpp"
-#include "UserVariables.hpp"
-#include "VariableType.hpp"
-#include "unistd.h" // gives 'access'
-#include <algorithm>
-#include <string>
-
-// Chombo namespace
-#include "UsingNamespace.H"
-
-class ChomboParameters
-{
-  public:
-    ChomboParameters(GRParmParse &pp)
-    {
-        read_params(pp);
-        check_params();
-    }
-
-    void read_params(GRParmParse &pp)
-    {
-        // must be before any pout() in the code to setPoutBaseName
-        read_filesystem_params(pp);
-
-        pp.load("verbosity", verbosity, 0);
-        // Grid setup
-        pp.load("max_spatial_derivative_order", max_spatial_derivative_order,
-                4);
-        pp.load("num_ghosts", num_ghosts,
-                (max_spatial_derivative_order == 6) ? 4 : 3);
-        pp.load("tag_buffer_size", tag_buffer_size, 3);
-        pp.load("grid_buffer_size", grid_buffer_size, 8);
-        pp.load("dt_multiplier", dt_multiplier, 0.25);
-        pp.load("fill_ratio", fill_ratio, 0.7);
-
-        // Periodicity and boundaries
-        boundary_params.read_params(pp);
-
-        // L's, N's and center
-        read_grid_params(pp);
-
-#ifdef CH_USE_HDF5
-        pp.load("ignore_checkpoint_name_mismatch",
-                ignore_checkpoint_name_mismatch, false);
-#endif
-
-        pp.load("max_level", max_level, 0);
-        // the reference ratio is hard coded to 2
-        // in principle it can be set to other values, but this is
-        // not recommended since we do not test GRChombo with other
-        // refinement ratios - use other values at your own risk
-        ref_ratios.resize(max_level + 1);
-        ref_ratios.assign(2);
-        pp.getarr("regrid_interval", regrid_interval, 0, max_level);
-        // Regridding on max_level does nothing but Chombo's AMR class
-        // expects this Vector to be of length max_level + 1
-        // so just set the final value to 0.
-        regrid_interval.resize(max_level + 1);
-        regrid_interval[max_level] = 0;
-
-        if (pp.contains("regrid_thresholds"))
-        {
-            pout() << "Using multiple regrid thresholds." << std::endl;
-            // As for regrid_interval, the last element is irrelevant
-            pp.getarr("regrid_thresholds", regrid_thresholds, 0, max_level);
-            regrid_thresholds.resize(max_level + 1);
-            regrid_thresholds[max_level] = regrid_thresholds[max_level - 1];
-        }
-        else
-        {
-            pout() << "Using single regrid threshold." << std::endl;
-            double regrid_threshold;
-            pp.load("regrid_threshold", regrid_threshold, 0.5);
-            regrid_thresholds = Vector<double>(max_level + 1, regrid_threshold);
-        }
-
-        // time stepping outputs and regrid data
-        pp.load("checkpoint_interval", checkpoint_interval, 1);
-        pp.load("plot_interval", plot_interval, 0);
[…]
