+++ Source/GRChombo/Tests/SphericalExtractionUniformTest/UserVariables.hpp
@@ -1,1 +1,0 @@
-../SphericalExtractionTest/UserVariables.hpp

[…]
