### 2024-06-24 0ac0bd5
Adjusted CI workflows to account for removal of GRChombo source
