### 2024-06-24 3f9e1fa
fixed typo in CI workflow
