### 2024-06-24 5d60e95
fixed incorrect environment variable in CI workflows
