### 2024-06-24 ea22bcd
Remove GRChombo source from GRBoondi source directory  (#15)
* Removed GRChombo source from source directory and added details in makefiles to set GRCHOMBO_SOURCE variable

* modified test makefiles to account for lack of GRChombo source

* Adjusted CI workflows to account for removal of GRChombo source

* fixed typo in CI workflow

* fixed incorrect environment variable in CI workflows
