+++ Source/GRChombo/Source/ApparentHorizonFinder/AHFunctions.hpp
@@ -1,389 +1,0 @@
-/* GRChombo
- * Copyright 2012 The GRChombo collaboration.
- * Please refer to LICENSE in GRChombo's root directory.
- */
-
-#ifndef _AHFUNCTIONS_HPP_
-#define _AHFUNCTIONS_HPP_
-
-// Chombo includes
-#include "CH_Timer.H"
-
-// Other includes
-#include "DimensionDefinitions.hpp" // make sure GR_SPACEDIM exists
-#include "TensorAlgebra.hpp"
-#include "UserVariables.hpp"
-#include <cmath> // sqrt
-
-#include "AHFunctionDefault.hpp"
-
-// Chombo namespace
-#include "UsingNamespace.H"
-
-/////////////////////////////////////////////////////////
-// Predefined optimization functions
-/////////////////////////////////////////////////////////
-
-struct ExpansionFunction : AHFunctionDefault
-{
-    //////////////////////////////////
-    // needed to calculate expansion:
-
-    // induced metric
-    Tensor<2, double> g;
-    Tensor<2, double> g_UU;
-    Tensor<3, double> dg;
-
-    // extrinsic curvature
-    Tensor<2, double> K;
-    double trK;
-
-    // target coordinate (e.g. radius)
-    double f;
-
-    // case of reduced Cartoon methods that have an extra metric component and
-    // require the coordinates to calculate the expansion
-#if GR_SPACEDIM != CH_SPACEDIM
-    // hd - higher dimensions
-    Tensor<1, double> coords; // cartesian coordinates
-    double g_hd;
-    Tensor<1, double> dg_hd;
-
-    double Kww; // not needed to calculate expansion, but may be needed for ut
-    ALWAYS_INLINE double get_metric_hd() const { return g_hd; }
-#endif
-    //////////////////////////////////
-
-    // only require variables up to Aij (chi, hij, K, Aij)
-    // this assumes c_Theta comes after the last of Aij
-    // (and like this the code is generic for 2D and 3D)
-    static ALWAYS_INLINE int vars_min() { return c_chi; }
-    static ALWAYS_INLINE int vars_max() { return c_Theta - 1; }
-    // Derivatives required only for the metric component (chi and hij)
-    // this assumes c_K comes after the last hij
-    // (and like this the code is generic for 2D and 3D)
-    static ALWAYS_INLINE int d1_vars_min() { return c_chi; }
-    static ALWAYS_INLINE int d1_vars_max() { return c_K - 1; }
-
-    ALWAYS_INLINE const Tensor<2, double> get_metric() const { return g; }
-    ALWAYS_INLINE const Tensor<2, double> get_inverse_metric() const
-    {
-        return g_UU;
-    }
-    ALWAYS_INLINE const Tensor<2, double> get_extrinsic_curvature() const
-    {
-        return K;
-    }
-
-    static int num_write_vars()
-    {
-        int num_components = CH_SPACEDIM * (CH_SPACEDIM + 1) / 2 * 2;
-#if GR_SPACEDIM != CH_SPACEDIM
-        num_components += 2; // hww and Kww
-#endif
-        return num_components;
-    }
-    static void write_headers(std::string *vec)
-    {
-        int el = 0;
-        for (int i = 0; i < CH_SPACEDIM; ++i)
-            for (int j = i; j < CH_SPACEDIM; ++j)
-                vec[el++] = "g" + std::to_string(i + 1) + std::to_string(j + 1);
-#if GR_SPACEDIM != CH_SPACEDIM
-        vec[el++] = "gww";
-#endif
-
-        for (int i = 0; i < CH_SPACEDIM; ++i)
[…]
