+++ Source/GRChombo/Source/GRChomboCore/VariableType.hpp
@@ -1,16 +1,0 @@
-/* GRChombo
- * Copyright 2012 The GRChombo collaboration.
- * Please refer to LICENSE in GRChombo's root directory.
- */
-
-#ifndef VARIABLETYPE_HPP
-#define VARIABLETYPE_HPP
-
-// enum for evolution or diagnostic variables
-enum class VariableType
-{
-    evolution,
-    diagnostic
-};
-
-#endif /* VARIABLETYPE_HPP */

[…]
