### 2024-06-24 #6888 @ShaunFell
@KAClough I've removed the GRChombo source code in https://github.com/ShaunFell/GRBoondi/pull/15 and updated the main branch
