### 2024-07-05 #6888 @Sbozzolo
Hi @ShaunFell, I agree with @KAClough. I think that it would be a net positive for everyone to work more closely with the GRChombo collaboration and re-use and share as much code as possible.

In any case, it should be very clear how GRChombo is used. As it stands, any update to GRChombo might end up breaking GRBoondi.

> Proca theories are quite complex so keeping GRBoondi separate from GRChombo allows me to be quite versatile in changing the code for the needs of higher-order theories in the generalized Proca landscape.

Choosing to move away from the released GRChombo has implications for the maintenance burden, so users should be provided with this information.
