### 2024-07-16 9d91fc2
Removed derivative class member from ProcaSquared class
