+++ Source/Diagnostics/ProcaSquared.hpp
@@ -19,7 +19,6 @@ template <class background_t> class ProcaSquared
 {
   protected:
     const double m_dx;
-    const FourthOrderDerivatives m_deriv;
     const std::array<double, CH_SPACEDIM> m_center;
     const background_t m_background;
 
@@ -31,7 +30,7 @@ template <class background_t> class ProcaSquared
   public:
     ProcaSquared(double a_dx, const std::array<double, CH_SPACEDIM> a_center,
                  const background_t a_background)
-        : m_dx(a_dx), m_deriv{m_dx}, m_center{a_center}, m_background{
+        : m_dx(a_dx), m_center{a_center}, m_background{
                                                              a_background} {};
 
     template <class data_t> void compute(Cell<data_t> current_cell) const
@@ -44,8 +43,6 @@ template <class background_t> class ProcaSquared
 
         // load variables from Chombo grid
         const auto matter_vars = current_cell.template load_vars<MatterVars>();
-        const auto matter_vars_d1 =
-            m_deriv.template diff1<MatterVars>(current_cell);
 
         // compute contravariant conformal spatial metric
         const auto gamma_UU{
