### 2024-07-19 1bf3898
added documentation for 2D plotter
