+++ PostProcessing/2DPlot.py
@@ -12,11 +12,12 @@ config = configparser.ConfigParser()
 config.read(sys.argv[1])
 
 
-
+# Extract config variables
 plotpath = config["Output"]["output_plot_path"]
 moviepath = config["Output"]["output_movie_path"]
 overwrite_plots = config["Output"].getboolean("overwrite_plots",0)
 
+#if the plot and movie paths dont exist, create them
 if not os.path.exists(plotpath):
 	os.mkdir(plotpath)
 if not os.path.exists(moviepath):
@@ -26,6 +27,7 @@ print("HDF5 file path: " + config["Header"]["hdf5_path"])
 print("Output Plots directory: " +plotpath)
 print("Output Movie Directory: " +moviepath)
 
+#set verbosity for printing
 def verbPrint(*objects):
 	if config["Header"].getint("verbosity",0):
 		print(*objects)
@@ -44,6 +46,7 @@ def setup_engine():
 		host = config["EngineConfig"].get("host", "localhost")
 		print("Running in parallel on host: ", host)
 	
+		# extract all the config options for the engine
 		num_procs = config["EngineConfig"].get("number_processes", 1)
 		num_nodes = config["EngineConfig"].get("number_nodes", 1)
 		partition_name = config["EngineConfig"].get("partition")
@@ -54,11 +57,22 @@ def setup_engine():
 
 		arg = ("-l", job_cmd, "-n", job_name, "-p", partition_name, "-np", num_procs, "-nn", num_nodes, "-t", time_limit, "-la", add_sub_args)
 
-		#Submit the job
+		#Submit the job using visit job scheduler functionality
 		OpenComputeEngine(host, arg)
 
+# This is where the real magic happens. This function creates a single slice plot.
 def setup_slice_plot(variableToPlot, plotbounds, setplotbounds) :
-	# annotation settings
+	"""
+	Create all the data for a single slice plot
+	"""
+
+	"""
+	For documentation on all the attributes below, see
+		https://visit-sphinx-github-user-manual.readthedocs.io/en/develop/python_scripting/functions.html
+	"""
+
+
+	# annotation settings, such as the information printed on the plot (user, database name, time, etc.)
 	AnnotationAtts = AnnotationAttributes()
 	axes3dvisible = config["AnnotationConfig"].getint('axes3Dvisible', 1)
 	verbPrint("Axes3d visible: ", str(axes3dvisible))
@@ -81,7 +95,7 @@ def setup_slice_plot(variableToPlot, plotbounds, setplotbounds) :
 	triadflag = config["AnnotationConfig"].getint('triadFlag', 0)
 	verbPrint("triadFlag: " , str(triadflag))
 
-
+	#Set the annotation attributes
 	AnnotationAtts.axes3D.visible = axes3dvisible
 	AnnotationAtts.userInfoFlag = userinfoflag
 	AnnotationAtts.databaseInfoFlag = databaseinfoflag
@@ -109,8 +123,8 @@ def setup_slice_plot(variableToPlot, plotbounds, setplotbounds) :
 	
 	# add pseudocolour plot
 	AddPlot("Pseudocolor", variableToPlot, 1, 1)
-	PseudocolorAtts = PseudocolorAttributes()
-	plotscaling = config["PlotConfig"].get("plotscaling", fallback = "Linear").lower()
+	PseudocolorAtts = PseudocolorAttributes() #Get the pseudocolor plot attributes instance
[…]
