### 2024-07-19 f77af64
added documentation for integral plotter and 3D plotter
