+++ PostProcessing/3DPlot.py
@@ -12,10 +12,11 @@ config = configparser.ConfigParser()
 config.read(sys.argv[1])
 
 
-
+# Extract config variables
 plotpath = config["Output"]["output_plot_path"]
 moviepath = config["Output"]["output_movie_path"]
 
+#if the plot and movie paths dont exist, create them
 if not os.path.exists(plotpath):
 	os.mkdir(plotpath)
 if not os.path.exists(moviepath):
@@ -25,6 +26,8 @@ print("HDF5 file path: " + config["Header"]["hdf5_path"])
 print("Output Plots directory: " +plotpath)
 print("Output Movie Directory: " +moviepath)
 
+
+#set verbosity for printing
 def verbPrint(*objects):
 	if config["Header"].getint("verbosity",0):
 		print(*objects)
@@ -74,7 +77,13 @@ def setup_slice_plot(variableToPlot, plotbounds, setplotbounds):
 	"""
 	This is where the real magic happens. Create all the data for the plot
 	"""
-	# annotation settings
+
+	"""
+	For documentation on all the attributes below, see
+		https://visit-sphinx-github-user-manual.readthedocs.io/en/develop/python_scripting/functions.html
+	"""
+
+	# annotation settings, such as the information printed on the plot (user, database name, time, etc.)
 	AnnotationAtts = AnnotationAttributes()
 
     ## Grab annotation settings from config file, with defaults
@@ -114,14 +123,15 @@ def setup_slice_plot(variableToPlot, plotbounds, setplotbounds):
 
 
 
-	#Add a volume plot
+	## Add a volume plot
 
 	AddPlot("Volume", variableToPlot,1,1)
 	VolumeAtts = VolumeAttributes()
 	plotscaling = config["VolumeConfig"].get("plotscaling", fallback = "Linear").title()
 	verbPrint("plot scaling: ",  plotscaling)
-	VolumeAtts.scaling = getattr(VolumeAtts, plotscaling) #Get the plot scaling attribute
 
+	#Set the attributes for the volume plot
+	VolumeAtts.scaling = getattr(VolumeAtts, plotscaling) #Get the plot scaling attribute
 	VolumeAtts.lightingFlag = config["VolumeConfig"].getboolean('lightingFlag', 1)
 	VolumeAtts.legendFlag =   config["VolumeConfig"].getboolean('legendFlag', 1)
 	VolumeAtts.opacityAttenuation = config["VolumeConfig"].getint('opacityAttenuation', 1)
@@ -140,6 +150,7 @@ def setup_slice_plot(variableToPlot, plotbounds, setplotbounds):
 	verbPrint("opacity ramp: ",  opacityramp)
 	VolumeAtts.freeformOpacity = opacityramp #set the tuple as the opacity ramp
 
+	#Determine the max and min values of the variable to use for coloring and opacity
 	VolumeAtts.useColorVarMin = config["VolumeConfig"].getboolean('useColorVarMin', 0)
 	VolumeAtts.colorVarMin = config["VolumeConfig"].getfloat('colorVarMin', 0)
 	VolumeAtts.useColorVarMax = config["VolumeConfig"].getboolean('useColorVarMax', 0)
@@ -149,6 +160,7 @@ def setup_slice_plot(variableToPlot, plotbounds, setplotbounds):
 	VolumeAtts.useOpacityVarMax = config["VolumeConfig"].getboolean('useOpacityVarMax', 0)
 	VolumeAtts.opacityVarMax = config["VolumeConfig"].getfloat('opacityVarMax', 0)
 
+	#Specify the type of rendering engine to use
 	rendertype = config["VolumeConfig"].get("rendererType", fallback = "default").title()
 	if rendertype == "Raycasting": rendertype = "RayCasting" 
[…]
