### 2024-07-19 8de1927
added elucidating comments to IntegralsPlot
