### 2024-07-19 fb6283a
added further documentation and fixed formatting
