### 2024-07-20 1241ada
added exceptions and checks to 3dplotter
