### 2024-07-25 9f14ecc
added error handling for post-processing routines
