### 2024-07-26 #6888 @ShaunFell
Hi @KAClough! I decided against making GRDzhadzha a dependency since that would add to the large dependency list and make downloading GRBoondi a bit more involved. I only use a very small portion of the GRDzhadzha codebase and I modify some of the files myself, so including it instead as a dependency is a bit unnecessary. I added this mention to the [GRBoondi wiki](https://github.com/ShaunFell/GRBoondi/wiki).

@Sbozzolo, I updated GRBoondi to make GRChombo a dependency, so updates to GRChombo now naturally flow into GRBoondi.
