### 2024-07-26 ShaunFell/GRBoondi#14 @ShaunFell
Hi @Sbozzolo! I've basically overhauled the wiki and included a detailed section on [using GRBoondi](https://github.com/ShaunFell/GRBoondi/wiki/Using-GRBoondi)
 * `csh` is a required program in [GRChombo](https://github.com/GRTLCollaboration/GRChombo/wiki/Prerequisites), but I suppose `bash` or other terminal environment would work
 * I added a brief summary to the parameters wiki page, but I think most of the parameters listed are quite important.
