+++ PostProcessing/2DPlot.py
@@ -27,376 +27,10 @@ print("HDF5 file path: " + config["Header"]["hdf5_path"])
 print("Output Plots directory: " +plotpath)
 print("Output Movie Directory: " +moviepath)
 
-#set verbosity for printing
-def verbPrint(*objects):
-	if config["Header"].getint("verbosity",0):
-		print(*objects)
+#Instantiate printer function that accepts a verbosity level
+verbPrint = VerbosityPrint(config["Header"].getint("verbosity",0))
 
-verbPrint("Verbosity: ", config["Header"].getint("verbosity",0))
-
-def setup_engine():
-	"""
-	Create the engine  for plotting
-	"""
-	useparallel = config["EngineConfig"].getboolean("use_parallel", 0)
-	
-	#if useparallel is disabled, default engine is launched
-	if useparallel:
-		host = config["EngineConfig"].get("host", "localhost")
-		print("Running in parallel on host: ", host)
-	
-		# extract all the config options for the engine
-		num_procs = config["EngineConfig"].get("number_processes", 1)
-		num_nodes = config["EngineConfig"].get("number_nodes", 1)
-		partition_name = config["EngineConfig"].get("partition")
-		job_cmd = config["EngineConfig"].get("job_cmd", "srun")
-		time_limit = config["EngineConfig"].get("time_limit", "01:00:00")
-		job_name = config["EngineConfig"].get("job_name", "3DPlot")
-		add_sub_args = config["EngineConfig"].get("additional_sub_args", "")
-
-		arg = ("-l", job_cmd, "-n", job_name, "-p", partition_name, "-np", num_procs, "-nn", num_nodes, "-t", time_limit, "-la", add_sub_args)
-
-		#Submit the job using visit job scheduler functionality
-		engine_status = OpenComputeEngine(host, arg)
-		if not engine_status:
-			raise RuntimeError("Could not start compute engine.")
-
-# This is where the real magic happens. This function creates a single slice plot.
-def setup_slice_plot(variableToPlot, plotbounds, setplotbounds) :
-	"""
-	Create all the data for a single slice plot
-	"""
-
-	"""
-	For documentation on all the attributes below, see
-		https://visit-sphinx-github-user-manual.readthedocs.io/en/develop/python_scripting/functions.html
-	"""
-
-
-	# annotation settings, such as the information printed on the plot (user, database name, time, etc.)
-	AnnotationAtts = AnnotationAttributes()
-	axes3dvisible = config["AnnotationConfig"].getint('axes3Dvisible', 1)
-	verbPrint("Axes3d visible: ", str(axes3dvisible))
-	userinfoflag = config["AnnotationConfig"].getint('userInfoFlag', 0)
-	verbPrint("userInfoFlag: ",str(userinfoflag))
-	databaseinfoflag = config["AnnotationConfig"].getint('databaseInfoFlag', 1)
-	verbPrint("databaseInfoFlag: " , str(databaseinfoflag))
-	timeinfoflag = config["AnnotationConfig"].getint('timeInfoFlag', 1)
-	verbPrint("timeInfoFlag: " , str(timeinfoflag))
-	legendinfoflag = config["AnnotationConfig"].getint('legendInfoFlag', 1)
-	verbPrint("legendInfoFlag: " , str(legendinfoflag))
-	axesarrayvisible = config["AnnotationConfig"].getint('axesArrayvisible', 1)
[…]
