+++ PostProcessing/Source/2D/__init__.py
@@ -1,1 +1,0 @@
-from Slice import *
