### 2024-08-11 e2ef424
added addition error checking
