+++ PostProcessing/Source/Common/PlotSetup.py
@@ -14,7 +14,7 @@ def Annotation_Setup(config):
 	"""Set the annotation settings. 
 
 	Args:
-		config (_type_): _description_
+		config (configparser.ConfigParser): instance of a ConfigParser class that holds the users parameters
 	"""
 
 	#initialize verbosity printing
@@ -25,24 +25,25 @@ def Annotation_Setup(config):
 
 	## Grab annotation settings from config file, with defaults
 	axes3dvisible = config["AnnotationConfig"].getint('axes3Dvisible', 1)
-	verbPrint("Axes3d visible: ", str(axes3dvisible))
 	userinfoflag = config["AnnotationConfig"].getint('userInfoFlag', 0)
-	verbPrint("userInfoFlag: ",str(userinfoflag))
 	databaseinfoflag = config["AnnotationConfig"].getint('databaseInfoFlag', 1)
-	verbPrint("databaseInfoFlag: " , str(databaseinfoflag))
 	timeinfoflag = config["AnnotationConfig"].getint('timeInfoFlag', 1)
-	verbPrint("timeInfoFlag: " , str(timeinfoflag))
 	legendinfoflag = config["AnnotationConfig"].getint('legendInfoFlag', 1)
-	verbPrint("legendInfoFlag: " , str(legendinfoflag))
 	axesarrayvisible = config["AnnotationConfig"].getint('axesArrayvisible', 1)
-	verbPrint('axesarrayvisible: ' , str(axesarrayvisible))
 	backgroundcolor = tuple(map(int,config["AnnotationConfig"].get("backgroundColor", "0 0 255").split()))
-	verbPrint("backgroundColor: " , backgroundcolor)
 	foregroundcolor = tuple(map(int, config["AnnotationConfig"].get("foregroundColor", "255 255 255").split()))
-	verbPrint("foregroundColor: " , foregroundcolor)
 	bboxflag = config["AnnotationConfig"].getint('bboxFlag', 1)
-	verbPrint("bboxFlag: " , str(bboxflag))
 	triadflag = config["AnnotationConfig"].getint('triadFlag', 0)
+
+	verbPrint("Axes3d visible: ", str(axes3dvisible))
+	verbPrint("userInfoFlag: ",str(userinfoflag))
+	verbPrint("databaseInfoFlag: " , str(databaseinfoflag))
+	verbPrint("timeInfoFlag: " , str(timeinfoflag))
+	verbPrint("legendInfoFlag: " , str(legendinfoflag))
+	verbPrint('axesarrayvisible: ' , str(axesarrayvisible))
+	verbPrint("backgroundColor: " , backgroundcolor)
+	verbPrint("foregroundColor: " , foregroundcolor)
+	verbPrint("bboxFlag: " , str(bboxflag))
 	verbPrint("triadFlag: " , str(triadflag))
 
 	#Set the annotation attributes
@@ -56,17 +57,34 @@ def Annotation_Setup(config):
 	AnnotationAtts.foregroundColor = foregroundcolor
 	AnnotationAtts.axes3D.bboxFlag = bboxflag
 	AnnotationAtts.axes3D.triadFlag = triadflag
-	visit.SetAnnotationAttributes(AnnotationAtts)
+
+	#save the annotation attributes
+	status = visit.SetAnnotationAttributes(AnnotationAtts)
+
+	if not status:
+		raise RuntimeError("Could not set annotation attributes. Aborting.")
 
 @require_visit
 def Pseudocolor_Setup(config, variableToPlot, setplotbounds, plotbounds):
+	"""Setup a pseudocolor plot
+
+	Args:
+		config (configparser.ConfigParser): instance of a ConfigParser class that holds the users parameters
+		variableToPlot (str): name of the variable to be plotted
[…]
