### 2024-08-11 96c14e2
fixed minor bugs and updated .gitignore
