### 2024-08-12 9e40a7c
removed checksum checks in post processing tests due to disagreement between different visit and python versions
