### 2024-08-12 59c9de9
added utility function to create movies with ffmpeg
