### 2024-08-12 d9854c2
debugging movie function
