### 2024-08-12 7deb256
fixed bug in ffmpeg movie function
