### 2024-08-12 6dac9eb
added additional output for exceptions in postprocessing main test function
