### 2024-08-12 b713f3f
updated failure exit codes for post processing tests
