### 2024-08-12 5084d45
updated default argument for volume plot render type
