+++ Tests/PostProcessing_test/3DPlotParams.ini
@@ -77,8 +77,8 @@ opacityVarMin = 0.00001
 useOpacityVarMax = 0
 opacityVarMax = 0.001
 
- ; default, raycasting, raycastingintegration, raycastingslivr, raycastingospray
-renderertype = default
+ ; serial, raycasting, raycastingintegration, raycastingslivr, raycastingospray
+renderertype = serial
 
  ; rasterization, kernelbased, trilinear
 sampling = rasterization
