### 2024-08-13 1d7ba67
updated postprocessing_test main function to account for additional passing exit codes
