### 2024-08-13 2f1f143
reverted default argument since not available in all versions of VisIt
