+++ PostProcessing/3DPlotParams.ini
@@ -80,8 +80,8 @@ opacityVarMin = 0.00001
 useOpacityVarMax = 0
 opacityVarMax = 0.001
 
- ; serial, raycasting, raycastingintegration, raycastingslivr, raycastingospray
-renderertype = serial
+ ; default, raycasting, raycastingintegration, raycastingslivr, raycastingospray
+renderertype = default
 
  ; rasterization, kernelbased, trilinear
 sampling = rasterization
