### 2024-08-13 c1b75aa
fixed incorrect print messages in test routines
