+++ Tests/DefaultBackground_test/Main_Tests.cpp
@@ -494,13 +494,13 @@ int test2()
     {
         std::cout << "The minimum convergence factor was "
                   << min_convergence_factor << std::endl;
-        std::cout << "Fixed Background test ........ passed" << std::endl;
+        std::cout << "Default Background test ........ passed" << std::endl;
     }
     else
     {
         std::cout << "The minimum convergence factor was "
                   << min_convergence_factor << std::endl;
-        std::cout << "Fixed Background test ........ failed!!!" << std::endl;
+        std::cout << "Default Background test ........ failed!!!" << std::endl;
     }
 
     return failed;
