+++ Tests/BoostedBH_test/Main_Tests.cpp
@@ -479,13 +479,13 @@ int test2()
     {
         std::cout << "The minimum convergence factor was "
                   << min_convergence_factor << std::endl;
-        std::cout << "Fixed Background test ........ passed" << std::endl;
+        std::cout << "Boosted BH test ........ passed" << std::endl;
     }
     else
     {
         std::cout << "The minimum convergence factor was "
                   << min_convergence_factor << std::endl;
-        std::cout << "Fixed Background test ........ failed!!!" << std::endl;
+        std::cout << "Boosted BH test ........ failed!!!" << std::endl;
     }
 
     return failed;
