### 2024-08-14 #6888 @KAClough
Hi @ShaunFell I see why you did it like this but having given this some more thought I do feel that GRDzhadzha should be a dependency. You currently don't use dynamical NR backgrounds (but I expect you may want to extend to that eventually), you use fixed backgrounds, which is very much a feature of GRDzhadzha and not GRChombo. Almost all your background metric code and the associated tests are directly taken from GRDzhadzha. We are currently developing a number of new background metrics that you would be able to take advantage of if you integrated with the GRDzhadzha code, whereas as it stands you would need to copy them over manually. Equally, we could benefit from the KerrdeSitter background if you separated it out and added it to GRDzhadzha rather than having it in this repo.

Realistically, your target market for this code is going to be GRChombo users, and if I were advising a student of mine who wanted to use non linear Proca-like matter, I would tell them to just copy over specific files from your repo to GRDzhadzha and fix any dependencies, rather than downloading your repo with so much duplication built into it. It would be quite annoying to have to check what, if any, changes you had made in all the similar looking files like LinearMomConservation.hpp for example.

In summary, I think you should remove all duplication, separate out general tools that could be useful to others using fixed backgrounds with different matter and add them to GRDzhadzha, and then focus this code repository on the specific Proca examples and diagnostics that you have developed so it is really clear where the boundary lies between the codes. This will also help if you do want to add NR backgrounds in future, as the background will always be provided by another code, and yours will sit clearly on top and interface with either.

If it helps we can discuss this in a zoom call some time, maybe with @Sbozzolo if he agrees?
