### 2024-08-15 #6888 @ShaunFell
Hi @KAClough! Thanks for your comments! I think a Zoom call would be great. I definitely see the excellent points in making GRDzhadzha a dependency and separating out things from my code, but I've made so many modifications to those individual files, it was just easier to modify each file than make GRDzhadzha an extra dependency and creating new classes that overloaded those functions. But we can discuss these things over a Zoom call!
