### 2024-08-18 #6888 @Sbozzolo
Hi @KAClough and @ShaunFell.

I cannot speak of the technical details and code provenance, but from the advertised capabilities, my impression is that the community would be best served if GRBoondi were a module of GRChombo (or GRDzhadzha) instead of a separated package. My primary concerns are fair credit attribution, maintenance burden, and alignment of development efforts. If this is not possible, I think that making GRDzhadzha a dependency is a choice that also depends on @ShaunFell’s future plans for GRBoondi.

Happy to have a conversation about this on Zoom!
