### 2024-08-19 #6888 @Sbozzolo
Assuming "tomorrow" is Tuesday 20 August, yes, 5 PM UK time (9 AM Pacific Time) would work for me.
