### 2024-08-19 #6888 @ShaunFell
That works for me! Im on Mountain time right now so that's not too early.
