### 2024-08-19 #6888 @KAClough
Ok, here is a zoom for tomorrow:

Topic: Katy Clough's Zoom Meeting
Time: Aug 20, 2024 05:00 PM London

Join Zoom Meeting
https://qmul-ac-uk.zoom.us/j/81262246232

Meeting ID: 812 6224 6232
