### 2024-08-25 11a7915
added .pyc files to .gitignore
