+++ Tests/PostProcessing_test/3DPlotParams.ini
@@ -43,7 +43,7 @@ additional_launch_args = --cpus-per-task=1 --mem=50gb
 plot_variables = Asquared
 
 [AnnotationConfig]
-axes3Dvisible = 0
+axes3Dvisible = 1
 userInfoFlag = 0
 databaseInfoFlag = 1
 legendInfoFlag = 1
