### 2024-09-03 #6888 @ShaunFell
Hi @xuanxu!

Yes, we all had a very productive meeting. Katy and I will be at a numerical relativity meeting next week and will discuss some finer details of the meeting points.

I just submitted my thesis last Thursday so haven't had any time to work on this submission, but I will starting in a few days again.
