+++ Tests/BoostedBH_test/Main_Tests.cpp
@@ -1,510 +1,0 @@
-/*
-GRBoondi
-
-Some of these tests were borrowed from the GRDzhadzha test suite. Please see
-https://github.com/GRTLCollaboration/GRDzhadzha/tree/main/Tests/KerrBHScalarTest
-for more details
-*/
-
-#ifdef _OPENMP
-#include <omp.h>
-#endif
-
-#define NUM_ERR_LIM 0.1
-
-// misc includes
-#include "AssignFixedBGtoBSSNVars.hpp"
-#include "ExcisionTest.hpp"
-
-// GRChombo includes
-#include "Cell.hpp"
-#include "Coordinates.hpp"
-#include "GammaCalculator.hpp"
-#include "MatterCCZ4.hpp"
-#include "MatterCCZ4RHS.hpp"
-#include "NewConstraints.hpp"
-
-// Chombo includes
-#include "BoxLoops.hpp"
-#include "Cell.hpp"
-#include "ComputePack.hpp"
-#include "DebuggingTools.hpp"
-#include "FArrayBox.H"
-#include "IntVect.H"
-#include "SetValue.hpp"
-#include "UsingNamespace.H"
-#include <fstream>
-#include <iomanip>
-#include <iostream>
-#include <sys/time.h>
-
-// background includes
-#include "ADMFixedBGVars.hpp"
-#include "BaseProcaFieldTest.hpp"
-#include "BoostedBH.hpp"
-#include "KerrSchildFixedBG.hpp"
-#include "MatterEvolution.hpp"
-#include "ProcaField.hpp"
-
-#include "BaseProcaField.hpp"
-#include "ProcaField.hpp"
-#include "ProcaFieldTest.hpp"
-
-// c++ includes
-#include <algorithm>
-#include <iomanip>
-#include <iostream>
-#include <string>
-
-std::string tab{"\t"};
-static const auto precision = std::setprecision(16);
-static const int printwidth = 25;
-
-// This struct should be a copy of the background param_t struct, with specified
-// values
-struct
-{
-    double mass = 1.0;
-    std::array<double, CH_SPACEDIM>
-        center; // automatically filled in during test2 eval
-    double velocity = 0.5;
-} BackgroundParams;
-
-int test2()
-{
-
-#ifdef _OPENMP
-    std::cout << " Number of threads: " << omp_get_max_threads() << std::endl;
-#endif
-
-    int failed{0};                    // flag for failed
-    const bool debug_plots_on{false}; // export data for plotting
-    const std::vector<int> resolutions{96, 192, 384}; // resolutions to run at
-    const int num_resolutions = 2; // how many of the resolutions to actually
-                                   // use
-
-    // vector of norms for convergence checking
-    std::array<std::array<double, NUM_VARS>, num_resolutions> error_norms;
-
-    for (int ires{0}; ires < num_resolutions; ++ires)
-    {
-        std::cout << std::endl << std::endl;
-        std::cout << "*******************" << std::endl;
-        std::cout << "Resolution: " << ires << std::endl;
-        std::cout << "*******************" << std::endl;
-
-        // fill current error norm array with zeroes
-        error_norms[ires].fill(0.0);
-
-        // setup the array boxes for various inputs and outputs
-        const int N_GRID{
-            resolutions[ires]}; // number of cells on each side of box
-
-        std::cout << "Creating grid (" << N_GRID << " x " << N_GRID << " x "
-                  << N_GRID << ")" << std::endl;
-        // setup boxes
-        Box box(IntVect(0, 0, 0), IntVect(N_GRID - 1, N_GRID - 1,
-                                          N_GRID - 1)); // The computational box
-        Box ghosted_box(
-            IntVect(-3, -3, -3),
-            IntVect(N_GRID + 2, N_GRID + 2,
-                    N_GRID +
[…]
