+++ PostProcessing/Source/OneD/LinePlot.py
@@ -82,3 +82,4 @@ def make_plots(config, data_obj, plot_variables, plotbounds = None,  linestyles 
         plt.legend()
         plt.ylabel("") #remove ylabel for combined plot
         save_pyplot_fig(config, "combined")
+        
