### 2024-09-04 0d1300e
changing variable structs to match GRDzhadzha name definitions
