+++ Tests/DefaultBackground_test/BaseProcaFieldTest.hpp
@@ -45,10 +45,10 @@ template <class background_t, class modification_t> class BaseProcaFieldTest
     BaseProcaFieldTest(background_t a_background)
         : m_background{a_background} {};
 
-    template <class data_t> using Vars = ADMProcaVars::MatterVars<data_t>;
+    template <class data_t> using Vars = ADMProcaVars::Vars<data_t>;
 
     template <class data_t>
-    using Diff2Vars = ADMProcaVars::Diff2MatterVars<data_t>;
+    using Diff2Vars = ADMProcaVars::Diff2Vars<data_t>;
 
     // we set G=0, so this doenst really matter, need it just for using
     // MatterCCZ4
