+++ Tests/L2_test/BaseProcaField.hpp
@@ -25,10 +25,10 @@ template <class background_t, class modification_t> class BaseProcaField
 
     template <class data_t> using MetricVars = ADMFixedBGVars::Vars<data_t>;
 
-    template <class data_t> using MatterVars = ADMProcaVars::MatterVars<data_t>;
+    template <class data_t> using Vars = ADMProcaVars::Vars<data_t>;
 
     template <class data_t>
-    using Diff2MatterVars = ADMProcaVars::Diff2MatterVars<data_t>;
+    using Diff2Vars = ADMProcaVars::Diff2Vars<data_t>;
 
     /* NOTE:
     We use the method of 'Curiously Recurring Template Pattern' to allow

[…]
