### 2024-09-04 edcbb92
Improved docstrings for many functions in post-processing
