+++ Source/Matter/BaseProcaFieldLevel.impl.hpp
@@ -46,7 +46,7 @@ void BaseProcaFieldLevel<background_t, proca_t>::prePlotLevel()
 
     // compute diagnostics on each cell of current level
     BoxLoops::loop(
-        make_compute_pack(Asquared, EM, z_field_diagnostic, background_init),
+        make_compute_pack(Asquared, EM, z_field_diagnostic),
         m_state_new, m_state_diagnostics, EXCLUDE_GHOST_CELLS);
 
     // Excise diagnostics within the excision zone.
