### 2024-09-07 8fabad8
changed 'if' conditionals to more readable form and replace list comprehension with 'filter' function
