### 2024-09-07 10aa7e4
Improvements: Post-Processing revamp (#18)
* added documentation for 2D plotter

* added documentation for integral plotter and 3D plotter

* added elucidating comments to IntegralsPlot

* added further documentation and fixed formatting

* added exceptions and checks to 3dplotter

* added error handling for post-processing routines

* Overhauled PostProcessing

* building test suite for PostProcessing

* Complete overhaul of PostProcessing routines. Moved all methods to python package-like directory and consolidate 2d and 3d main scripts to single file labelled VisitPlot.py
Added new test sui
[…]
