+++ PostProcessing/Source/Common/Utils.py
@@ -1,0 +1,379 @@
+## GRBoondi 2024
+## Copyright 2024, Shaun Fell
+## Please refer to LICENSE in GRBoondi's root directory
+
+import os, glob, sys
+import Source
+import configparser
+
+## Note:
+# we must put user-installed modules inside the functions that use them since VisIt doesn't support additional modules
+# and it's a pain to set it up for that. For simplicity, we put them there.
+
+## flag for VisIt
+try:
+    import visit
+    __visit_imported = True
+except:
+    __visit_imported = False
+
+## flag for Python
+try:
+    import matplotlib
+    import pandas
+    __extra_python = True
+except:
+    __extra_python = False
+
+#decorator for functions that require VisIt
+def require_visit(fn):
+    def wrapper(*args, **kwargs):
+        if not __visit_imported:
+            raise RuntimeError("VisIt not found. Aborting.")
+        import visit
+        return fn(*args, **kwargs)
+    return wrapper
+
+#decorator for functions that require non-standard python packages, e.g. matplotlib and pandas
+def require_python(fn):
+    def wrapper(*args, **kwargs):
+        if not __extra_python:
+            raise RuntimeError("Non-standard python packages not found. Aborting.")
+        import matplotlib.pyplot
+        import pandas
+        return fn(*args, **kwargs)
+    return wrapper
+
+
+def get_hdf5_file_list(config):
+    """Find all the hdf5 files and return them as list of absolute path strings
+
+    Args:
+        config (configparser.ConfigParser): instance of a ConfigParser class that holds the users parameters
+
+    Raises:
+        FileNotFoundError: Error raised if hdf5 path isn't found
+        FileNotFoundError: Error raised if no plot files were found
+
+    Returns:
+        list: list of plot file names
+    """
+    hdf5_path = config["Header"]["hdf5_path"]
+    filenames = config["Header"]["plot_header"] + "*.hdf5"
+
+    if not os.path.exists(hdf5_path):
+        raise FileNotFoundError("HDF5 path not found: " + hdf5_path)
+
+    #get list of file using regex 
+    hdf5files = glob.glob(os.path.join(hdf5_path, filenames))
+    hdf5files.sort() # sort the files by number
+
+    #check if files exist
+    if not hdf5files:
+        raise FileNotFoundError("No HDF5 files found in " + config["Header"]["hdf5_path"])
+
+    return hdf5files
+
+def get_config(arg_list):
+    """safely get config object from command line arguments
+
+    Args:
+        arg_list (list): list of arguments from command line
+
+    Raises:
+        RuntimeError: please specify a parameter file
+
+    Returns:
+        configparser.ConfigParser: instance of a ConfigParser class that holds the users parameters
+    """
+
+    if len(arg_list) < 2:
+        raise RuntimeError("Please specify a parameter file")
+    
+    config = configparser.ConfigParser()
+    config.read(arg_list[1])
+
+    if __visit_imported: #If we're running VisIt, then these are the parameter sections we expect
+        mandatory_visit_headers = ["Header", "EngineConfig", "VariableData", 
+                            "AnnotationConfig", "PlotConfig", 
+                            "ViewConfig", "Output"]
+    elif __extra_python: #If we're running via python, then these are the parameter sections expect
+        mandatory_visit_headers = ["Header", "VariableData", "PlotConfig", "Output"]
+
+
+    loaded_config_sections = config.sections()  
+
+    for header in mandatory_visit_headers:
+        if header not in loaded_config_sections:
+            raise configparser.NoSectionError("Parameter file error. Missing section: " + header)
+
+    return config
+
+def create_output_dirs(config):
[…]
