### 2024-09-13 ShaunFell/GRBoondi#14 @ShaunFell
Hi @Sbozzolo !

To keep this issue updated, i've added further edits to the wiki, including documenting upcoming changes suggested by Katy (adding GRDzhadzha as a dependency, PR #20 ) and estimating resource requirements.
