+++ Tests/L2_test/BaseProcaField.hpp
@@ -3,7 +3,7 @@
 
 #include "ADMFixedBGVars.hpp"         //For metric variables
 #include "ADMProcaVars.hpp"           //For matter variables
-#include "DefaultBackground.hpp"      //Minkowski background as default
+#include "Minkowski.hpp"      //Minkowski background as default
 #include "FourthOrderDerivatives.hpp" //For calculating derivatives
 #include "Tensor.hpp"                 //For performing tensorial operations
 #include "TensorAlgebra.hpp"
