+++ Tests/DefaultBackground_test/ProcaFieldTest.hpp
@@ -7,14 +7,14 @@ This class adds the simplest L2 lagrangian to the base equations of motion
 #include "ADMFixedBGVars.hpp"
 #include "ADMProcaVars.hpp"
 #include "BaseProcaField.hpp"
-#include "DefaultBackground.hpp"
+#include "Minkowski.hpp"
 #include "DefaultG.hpp"
 #include "L2_simp.hpp"
 
 // Note: base class BaseProcaField uses CRTP, so pass ProcaField itself as
 // template argument
 class ProcaFieldTest
-    : public BaseProcaFieldTest<DefaultBackground, ProcaFieldTest>
+    : public BaseProcaFieldTest<Minkowski, ProcaFieldTest>
 {
 
   public:
@@ -36,13 +36,13 @@ class ProcaFieldTest
         double vector_damping;
     };
 
-    DefaultBackground m_background;
+    Minkowski m_background;
     params_t m_params;
     L2_t m_L2;
     DefaultG m_G2;
 
-    ProcaFieldTest(DefaultBackground a_background, params_t a_params)
-        : BaseProcaFieldTest<DefaultBackground, ProcaFieldTest>(a_background),
+    ProcaFieldTest(Minkowski a_background, params_t a_params)
+        : BaseProcaFieldTest<Minkowski, ProcaFieldTest>(a_background),
           m_background(a_background), m_params(a_params)
     {
         // set up the L2 lagrangian
