### 2024-10-02 c0fc5b3
renamed all instances of KerrSchildNew to KerrSchild Renamed DefaultBackground file to Minkowski
