+++ Tests/L2_test/SimulationParameters.hpp
@@ -1,7 +1,7 @@
 #ifndef SIMULATIONPARAMETERS_HPP_
 #define SIMULATIONPARAMETERS_HPP_
 
-#include "KerrSchildNew.hpp"
+#include "KerrSchild.hpp"
 #include "L2_simp.hpp"
 #include "ProcaField.hpp"
 #include "ProcaSimulationParameters.hpp"
@@ -57,7 +57,7 @@ class SimulationParameters : public ProcaSimulationParameters
     }
 
     // parameters of kerr bh
-    KerrSchildNew::params_t background_params;
+    KerrSchild::params_t background_params;
 
     // initial conditions parameters
     //        Initial_Proca_Conditions::params_t initial_conditions_params;
