+++ Tests/KerrdeSitter_test/BaseProcaFieldTest.hpp
@@ -20,8 +20,8 @@ classes. e.g. see ProcaKerrBH Example
 #include "ADMFixedBGVars.hpp" //For metric variables
 #include "ADMProcaVars.hpp"   //For matter variables
 #include "CCZ4RHS.hpp"
-#include "Minkowski.hpp"      //Minkowski background as default
 #include "FourthOrderDerivatives.hpp" //For calculating derivatives
+#include "Minkowski.hpp"              //Minkowski background as default
 #include "Tensor.hpp"                 //For performing tensorial operations
 #include "TensorAlgebra.hpp"
 #include "UserVariables.hpp" //For user-defined variables (e.g. see EMKerrBH)
