+++ Source/Diagnostics/DampingField.hpp
@@ -29,7 +29,7 @@ class DampingFieldDiagnostic
                                                             // matter variables
 
   public:
-    DampingFieldDiagnostic() {}; // explicit default constructor
+    DampingFieldDiagnostic(){}; // explicit default constructor
 
     template <class data_t> void compute(Cell<data_t> current_cell) const
     {
