### 2024-10-02 4f6a7d1
Improvements: Add GRDzhadzha as dependency (#20)
* Removed redundant code and added GRDzhadzha as dependency

* changing variable structs to match GRDzhadzha name definitions

* Set GRDzhadzha as dependency and debugged examples/tests

* update CI workflows to include GRDzhadzha checkout

* clang format

* Removed redundant code
Renamed Matter folder to ProcaFields
Moved FluxExtraction code to Diagnostics folder
Renamed DefaultBackground class to Minkowski

* Renamed Matter folder in GNUMakeFiles to ProcaFields

* renamed all instances of KerrSchildNew to KerrSchild
Renamed DefaultBackground f
[…]
