### 2024-10-16 #6888 @KAClough
Some comments on the code itself and the wiki:

1. There are some `output_path=` in the `params.txt` files that give your own file system path, these should be set to "", and naively running them gives an error that will confuse new users (who haven't made the same mistake themselves yet!)

2. The wiki could do with another thorough read over - mostly it is fine but there are a few places where references to GRDzhadzha don't completely make sense in the updated version - i.e. they still don't imply it is a dependency, some details of the background class tests aren't needed any more as they were removed, "where is everything" contains some out of date info etc.

3. Expected Required Computing Resources - these worked and I got comparable performance but it would be nice if they corresponded exactly to the `params_cheap.txt` file that is in the repository - this way it is easier for people to run directly as a check!

Apart from these things I am happy, so I have ticked all the boxes but one. I will review the paper separately.
