### 2024-10-16 #6888 @KAClough
Comments on the paper:

- In my view the intro needs to be simplified to meet the needs of a non specialist audience - you need to start with a clear description for a lay scientist of what Proca fields are and how they arise in physics - you could then give the general equation of motion, which would make it clear the equation that is being solved. You should also give a brief discussion of NR versus fixed backgrounds, again at a basic level, so that the distinction you make later is clear.

- Again you need to update the paper for the GRDzhadzha dependency change.

- Another recent addition to other NR codes is ExoGRype by Baojiu Li and Han Zhang, which could be added to the citations.

- I am not sure your LOC comparison is very fair! You already use the underlying code of GRChombo, so naturally you have fewer lines of code! I think it is sufficient justification to say that you have created something that is tailored to the particular problem, so that it requires the user to look less at the underlying code. On the other hand, you say in the wiki that it is not a black box... and in many places you point to the GRChombo documentation, so that's not entirely consistent either. It is probably best to say that this is a collection of specialised tools for Generalised Proca, that relies on existing tools because this allows you to gain from any updates/bug fixes/speed ups that are implemented there.

- You say that you can use the output data as input for GRChombo, but that's not entirely straightforward - you would have to output an hdf5 file in the right format with the metric vars added - do you have code that does this? Even if you do, you would not have constraint satisfying data, so it wouldn't be possible to use it as direct input, you would need to put it into the initial condition solver (which is becoming public soon!).

- Do you use the 6th order stencils? The wiki says 4th but the paper says both 4th and 6th.

Again I think you are almost there so I have ticked all but one box!
