### 2024-10-17 78d4bf8
removed specification of output_path in all example directories and replaced with blank string. changed cheap parameter files in ProcaKerrBH example to mirror the expected computing resources on the wiki
