+++ Examples/ProcaKerrBH/params/params_cheap.txt
@@ -6,26 +6,26 @@
 verbosity = 1
 
 # location / naming of output files
-output_path = "/mnt/DataDrive/Data_Store/GRBoondi/ProcaKerrBH"
+output_path=""
 
 
 #filename prefixes for checkpoint files and plot files
-chk_prefix = "GeneralizedProca_"
-plot_prefix = "test_dataset"
+chk_prefix = "test_dataset_"
+plot_prefix = "test_datasetp_"
 
 #file to restart the simulation from
 #restart_file = GeneralizedProca_002200.3d.hdf5
 run_initial_posttimestep = 1
 
 # HDF5files are written every dt = L/N*dt_multiplier*checkpoint_interval
-checkpoint_interval = 100
+checkpoint_interval = 10000
 # set to 0 to turn off plot files (except at t=0 and t=stop_time)
 # set to -1 to never ever print plotfiles
-plot_interval = 1
+plot_interval = -1
 
 #plot variable names must match those in the DiagnosticVariables::DiagnosticVariables namespace
-num_plot_vars = 1
-plot_vars = Asquared
+num_plot_vars = 0
+plot_vars =
 # subpaths - specific directories for hdf5, pout, extraction data
 # (these are created at runtime)
 hdf5_subpath = "hdf5"
@@ -50,10 +50,10 @@ activate_tagging_diagnostic = 0
 ################################################# AMR levels
 ##################################################
 
-max_level = 3
+max_level = 4
 
 #Since we use fixed grid tagging, we dont ever need to regrid
-regrid_interval = 0 0 0
+regrid_interval = 0 0 0 0
 
 ################################################## Initial Data parameters
 ##################################################
@@ -84,8 +84,8 @@ bh_evolution = 0
 ################################################## Grid parameters
 ##################################################
 
-N_full = 32
-L_full = 32
+N_full = 64
+L_full = 64
 
 # Max and min box sizes
 max_box_size = 16
@@ -178,7 +178,7 @@ diagnostic_excision_vars = rho rhoE rhoJ rhoLinMom sourceLinMom
 ##################################################
 
 activate_extraction = 0 # extraction of fluxes
-activate_integration = 1 # integration of densities
+activate_integration = 0 # integration of densities
 
 num_extraction_radii = 2
 extraction_radii = 1.6 14
