### 2024-10-17 413cbeb
remote output_path specification and replaced with empty string
