+++ Examples/ProcaKerrdeSitter/params/params_cheap.txt
@@ -7,7 +7,7 @@
 verbosity = 1
 
 # location / naming of output files
-output_path = "/mnt/DataDrive/Data_Store/GRBoondi/ProcaKerrdeSitter"
+output_path=""
 
 #filename prefixes for checkpoint files and plot files
 chk_prefix = "GeneralizedProca_"
