+++ Examples/ProcaKerrdeSitter/params/params_highres_re.txt
@@ -7,7 +7,7 @@ restart_file=GeneralizedProca_001000.3d.hdf5
 verbosity = 0
 
 # location / naming of output files
-output_path = "/pfs/work7/workspace/scratch/hd_pb293-WS_GRBoondi/testing/re_dev_highres_GRMilijun_ProcaKdS_23530622"
+output_path=""
 
 #filename prefixes for checkpoint files and plot files
 chk_prefix = "GeneralizedProca_"
