### 2024-10-17 dd2087f
updated paper contents according to JOSS reviewer comments
