+++ paper/paper.bib
@@ -598,3 +598,13 @@
 	archiveprefix = {arXiv},
 	primaryclass = {astro-ph.CO}
 }
+
+@misc{zhang2024exagrypenumericalgeneralrelativity,
+      title={ExaGRyPE: Numerical General Relativity Solvers Based upon the Hyperbolic PDEs Solver Engine ExaHyPE}, 
+      author={Han Zhang and Baojiu Li and Tobias Weinzierl and Cristian Barrera-Hinojosa},
+      year={2024},
+      eprint={2406.11626},
+      archivePrefix={arXiv},
+      primaryClass={gr-qc},
+      url={https://arxiv.org/abs/2406.11626}, 
+}
