### 2024-10-17 5935792
update to paper contents from JOSS reviewer comments
