+++ paper/paper.md
@@ -25,25 +25,25 @@ bibliography: paper.bib
 
 # Summary
 
-Proca theories are a simple massive extension of the typical electromagnetic field. Such a simple modification of the underlying theory can have tremendous phenomenological implications, particularly in the case of early-universe high-energy physics, structure formation, and various dark matter models. On the theoretical side, they arise naturally in certain string theories and axionic interactions. Phenomenologically, Proca fields can be produced in the early universe from quantum fluctuations. This leads the Proca field to be an excellent candidate for the dark matter particle. Naturally, generalizations of the Proca field can lead to even richer astrophyical implications. These extensions provide possible solutions to many challenges across a wide range of astrophysical phenomena.
+Proca theories are a simple massive extension of the typical electromagnetic field. Such a simple modification of the underlying theory can have tremendous phenomenological implications, particularly in the case of early-universe high-energy physics, structure formation, and various dark matter models. On the theoretical side, they arise naturally in certain string theories and axionic interactions. Phenomenologically, Proca fields can be produced in the early universe from quantum fluctuations. This leads the Proca field to be an excellent candidate for the dark matter particle. Naturally, generalizations of the Proca field can lead to even richer astrophyical implications, providing possible solutions to many challenges across a wide range of astrophysical phenomena.
 
 ## Generalized Proca 
 
-Generalized Proca theories [@Heisenberg_2019;@Heisenberg_2014] provide a rich landscape to search for solutions to deep, fundamental questions, such as the nature of dark energy and dark matter. Moreover, strong gravity regimes, like those surrounding dense, compact astrophysical entities, offer novel avenues to explore fundamental fields. To effectively probe these domains, precise models are imperative for sifting through vast quantities of data. Crafting concrete models for the entirety of Generalized Proca theories represents a formidable endeavor, typically relying on numerical methods. These numerical methods are usually written to be case dependent, especially for a particular background. This makes it difficult to generalize the computational code to account for higher-order couplings or different types of backgrounds. The difficulty is especially amplified when considering the full landscape of Generalized Proca theories and beyond [@Heisenberg:2016eld]. Already at the level of the Lagrangian, the equations are immense,
+Generalized Proca theories [@Heisenberg_2019;@Heisenberg_2014] provide a rich landscape to search for solutions to deep, fundamental questions, such as the nature of dark energy and dark matter. Moreover, strong gravity regimes, like those surrounding dense, compact astrophysical entities, offer novel avenues to explore fundamental fields. To effectively probe these domains, precise models are imperative for sifting through vast quantities of data. Crafting concrete models for the entirety of generalized Proca theories represents a formidable endeavor, typically relying on numerical methods. These numerical methods are usually written to be case dependent, especially for a particular background. This makes it difficult to generalize the computational code to account for higher-order couplings or different types of backgrounds. The difficulty is especially amplified when considering the full landscape of generalized Proca theories and beyond [@Heisenberg:2016eld]. Already at the level of the Lagrangian, the equations are immense,
 $$
-	\mathcal{L}_{\text{g.P.}} = \sqrt{-g} \left(-\frac{1}{4}F_{\mu \nu}F^{\mu \nu} + \displaystyle\sum_{n=2}^{6} \alpha_n \mathcal{L}_n \right) \;.
+	\mathcal{L}_{\text{g.P.}} = \sqrt{-g} \left(-\frac{1}{4}F_{\mu \nu}F^{\mu \nu} + \displaystyle\sum_{n=2}^{6} \alpha_n \mathcal{L}_n \right) \;,
 $$
[…]
