### 2024-10-17 #6888 @ShaunFell
>     1. There are some `output_path=` in the `params.txt` files that give your own file system path, these should be set to "", and naively running them gives an error that will confuse new users (who haven't made the same mistake themselves yet!)

fixed! Forgot to remove those, thanks for pointing that out.

>     2. The wiki could do with another thorough read over - mostly it is fine but there are a few places where references to GRDzhadzha don't completely make sense in the updated version - i.e. they still don't imply it is a dependency, some details of the background class tests aren't needed any more as they were removed, "where is everything" contains some out of date info etc.

I went back through the wiki and updated the references to GRDzhadzha, all the ones I could find.

>     3. Expected Required Computing Resources - these worked and I got comparable performance but it would be nice if they corresponded exactly to the `params_cheap.txt` file that is in the repository - this way it is easier for people to run directly as a check!

done!

>     * In my view the intro needs to be simplified to meet the needs of a non specialist audience - you need to start with a clear description for a lay scientist of what Proca fields are and how they arise in physics - you could then give the general equation of motion, which would make it clear the equation that is being solved. You should also give a brief discussion of NR versus fixed backgrounds, again at a basic level, so that the distinction you make later is clear.

I did my best to simplify it. I added an introduction to standard Proca and the generalized Proca lagrangian, citing Lavinia's paper for the specific equations. In my opinion, the last paragraph of the summary section (right above the Statement of needs) already gives a clear distinction between solving full Einstein equations and the fixed background approach. What do you think?

>     * Again you need to update the paper for the GRDzhadzha dependency change.

updated

>     * Another recent addition to other NR codes is ExoGRype by Baojiu Li and Han Zhang, which could be added to the citations.

cool! I added a citation for this

>     * I am not sure your LOC comparison is very fair! You already use the underlying code of GRChombo, so naturally you have fewer lines of code! I think it is sufficient justification to say that you have created something that is tailored to the particular problem, so that it requires the user to look less at the underlying code. On the other hand, you say in the wiki that it is not a black box... and in many places you point to the GRChombo documentation, so that's not entirely consistent either. It is probably best to say that this is a collection of specialised tools for Generalised Proca, that relies on existing tools because this allows you to gain from any updates/bug fixes/speed ups that are implemented there.

This is a fair remark. In my head, I was thinking more about setting up an example directory: GRChombo typically requires the user to write a lot more code since its a much more powerful codebase and requires more input. I removed this comparision and just talked about GRBoondi being a collection of tools for this specialized problem.

>     * You say that you can use the output data as input for GRChombo, but that's not entirely straightforward - you would have to output an hdf5 file in the right format with the metric vars added - do you have code that does this? Even if you do, you would not have constraint satisfying data, so it wouldn't be possible to use it as direct input, you would need to put it into the initial condition solver (which is becoming public soon!).

I don't think I say GRBoondi outputs can be directly input into GRChombo, rather it can be easily ported (maybe some quantities require transformations or something), but since the checkpoint files are the same, the two codebases can easily talk with each other.

>     * Do you use the 6th order stencils? The wiki says 4th but the paper says both 4th and 6th.

not really, but can be used since it comes from GRChombo. I added a mention to this.

> Again I think you are almost there so I have ticked all but one box!

:D
