+++ paper/paper.md
@@ -33,7 +33,7 @@ Generalized Proca theories [@Heisenberg_2019;@Heisenberg_2014] provide a rich la
 $$
 	\mathcal{L}_{\text{g.P.}} = \sqrt{-g} \left(-\frac{1}{4}F_{\mu \nu}F^{\mu \nu} + \displaystyle\sum_{n=2}^{6} \alpha_n \mathcal{L}_n \right) \;,
 $$
-where each sub-Lagranian $\mathcal{L}_n$ contains terms of different orders in the derivatives of the Proca field [@Heisenberg_2019;@Heisenberg_2014]. Already at $\mathcal{L}_2$, the equations of motion can be immense. Solving these equations of motion analytically very quickly becomes intractable.
+where each sub-Lagranian $\mathcal{L}_n$ contains terms of different orders in the derivatives of the Proca field [@Heisenberg_2019;@Heisenberg_2014]. Even including solely $\mathcal{L}_2$, the equations of motion can be extremely cumbersome. Solving these equations of motion analytically very quickly becomes intractable.
 
 GRBoondi provides a unified interface for computing the evolution of any generalized Proca model on an arbitrary background. Given a specific background with known expressions for the metric variables and initial data for the Proca field, GRBoondi numerically computes the time evolution of the Proca field, given user-specified generalized Proca equations of motion. While the simultaneous evolution of the metric and matter fields offers the most comprehensive depiction of their temporal progression, there are instances where the density of the Proca field is negligible in comparison to the background curvature. In such cases, fixed background evolution routines serve as an excellent approximation to the complete mutual evolution. A core feature of GRBoondi is that it simplifies many of the boiler-plate code required to begin a simulation from scratch. The only additions a user needs to input are the initial conditions, the modifications to the equations of motion (EOM), and the background functions. GRBoondi will automatically compute various diagnostic quantities and plot files. Moreover, GRBoondi is incredibly modular and modifications to any of the in-built functions is effortless.
 
