+++ paper/paper.md
@@ -41,7 +41,7 @@ The fixed background approximation provides a huge speedup compared to the full 
 
 # Statement of need
 
-In practice, any NR software library can be used to evolve generalized Proca theories. Prominent examples of NR codes encompass the extensive Einstein toolkit [@EinsteinToolkit:2023_11] and its related Cactus framework [@Goodale:2002a], Kranc [@Kranc:web], LEAN [@Sperhake:2006cy], and Canuda [@Canuda:zenodo]. Other libraries worth mentioning are the non-public BAM [@Bruegmann:2006ulg], AMSS-NCKU [@Galaviz:2010mx], PAMR [@East:2011aa] and HAD [@Neilsen:2007ua]. The non-exhaustive list can be expanded with SPeC [@Pfeiffer:2002wt], which is a pseudo-spectral code that uses generalized harmonic coordinates. Similarly, SpECTRE [@deppe_nils_2021_4734670;@Kidder:2016hev;@Cao:2018vhw] uses the Galerkin methods. NRPy [@Ruchlin:2017com] is a python library that aims for use on non-high performance computing clusters. There are also cosmological simulation codes like CosmoGRaPH [@Mertens:2015ttp] and GRAMSES [@Barrera-Hinojosa:2019mzo]. Simflowny [@Palenzuela:2018sly] is a magneto-hydrodynamic simulation software. GRAthena++ [@Daszuta:2021ecf] uses the oct-tree AMR for maximum scaling. ExaGRyPE is a numerical relativity program leveraging the ExaHyPRE PDE solver [@zhang2024exagrypenumericalgeneralrelativity]. GRDzhadzha [@Aurrekoetxea2024] is another fixed background code, based off the GRChombo [@Andrade2021] framework, which GRBoondi utilizes.
+In practice, any NR software library can be used to evolve generalized Proca theories. Prominent examples of NR codes encompass the extensive Einstein toolkit [@EinsteinToolkit:2023_11] and its related Cactus framework [@Goodale:2002a], Kranc [@Kranc:web], LEAN [@Sperhake:2006cy], and Canuda [@Canuda:zenodo]. Other libraries worth mentioning are the non-public BAM [@Bruegmann:2006ulg], AMSS-NCKU [@Galaviz:2010mx], PAMR [@East:2011aa] and HAD [@Neilsen:2007ua]. The non-exhaustive list can be expanded with SPeC [@Pfeiffer:2002wt], which is a pseudo-spectral code that uses generalized harmonic coordinates. Similarly, SpECTRE [@deppe_nils_2021_4734670;@Kidder:2016hev;@Cao:2018vhw] uses the Galerkin methods. NRPy [@Ruchlin:2017com] is a Python library that aims for use on non-high performance computing clusters. There are also cosmological simulation codes like CosmoGRaPH [@Mertens:2015ttp] and GRAMSES [@Barrera-Hinojosa:2019mzo]. Simflowny [@Palenzuela:2018sly] is a magneto-hydrodynamic simulation software. GRAthena++ [@Daszuta:2021ecf] uses the oct-tree AMR for maximum scaling. ExaGRyPE is a numerical relativity program leveraging the ExaHyPRE PDE solver [@zhang2024exagrypenumericalgeneralrelativity]. GRDzhadzha [@Aurrekoetxea2024] is another fixed background code, based off the GRChombo [@Andrade2021] framework, which GRBoondi utilizes.
 
[…]
