+++ paper/paper.bib
@@ -605,6 +605,7 @@
       year={2024},
       eprint={2406.11626},
       archivePrefix={arXiv},
+	  doi = {10.2139/ssrn.4886174},
       primaryClass={gr-qc},
       url={https://arxiv.org/abs/2406.11626}, 
 }
