### 2024-10-23 #6888 @xuanxu
@ShaunFell It is not mandatory, but I'd suggest you edit the title at Zenodo to be the same as the paper (*GRBoondi: A code for evolving Generalized Proca theories on arbitrary backgrounds*)
