### 2024-10-23 #6888 @ShaunFell
> @ShaunFell It is not mandatory, but I'd suggest you edit the title at Zenodo to be the same as the paper (_GRBoondi: A code for evolving Generalized Proca theories on arbitrary backgrounds_)

Ah, good point. Changed!

Thank you @KAClough , @Sbozzolo , and @xuanxu for your efforts, reviewing my code, and helping me improve upon it. This has been a great experience!
