### 2024-10-29 #6888 @warrickball
Hi @ShaunFell, this has bubbled near the top of my list. My main concern is that the paper is very long by JOSS's standards. As indicated [in our docs](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain),

> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

If we can at least get "closer" to 1000 words (I'll certainly allow 1250 and _maybe_ a bit more), I can have a closer look at the remaining text.

My main feeling is that the "Key Features of GRBoondi" of section can be more or less removed or at least significantly shortened. For example, that you can use an arbitrary background has already been emphasised repeatedly in the text (e.g. second paragraph of "Generalised Proca" & much of the "Statement of Need"). Things like diagnostics, post-processing tools and boundary conditions are likely to change (or at least be extended) and are better off in docs. But if you can cut the paper by ~a quarter by other means, that's fine too.

On a more technical note, the citation to the arXiv version of Coates & Ramazanoğlu (2022) can be replaced with the [published version](https://ui.adsabs.harvard.edu/abs/2022PhRvL.129o1103C).
