### 2024-10-30 2051e2c
reduction of paper workcount, per JOSS review comments
