+++ paper/paper.bib
@@ -79,18 +79,20 @@
 	url          = {http://dx.doi.org/10.1103/PhysRevLett.129.151102}
 }
 @article{Coates_2022,
-	title        = {Intrinsic Pathology of Self-Interacting Vector Fields},
-	author       = {Coates, Andrew and Ramazano\u{g}lu, Fethi M.},
-	year         = 2022,
-	month        = oct,
-	journal      = {Physical Review Letters},
-	publisher    = {American Physical Society (APS)},
-	volume       = 129,
-	number       = 15,
-	doi          = {10.1103/physrevlett.129.151103},
-	issn         = {1079-7114},
-	url          = {http://dx.doi.org/10.1103/PhysRevLett.129.151103}
+  title = {Intrinsic Pathology of Self-Interacting Vector Fields},
+  author = {Coates, Andrew and Ramazano\ifmmode \breve{g}\else \u{g}\fi{}lu, Fethi M.},
+  journal = {Phys. Rev. Lett.},
+  volume = {129},
+  issue = {15},
+  pages = {151103},
+  numpages = {6},
+  year = {2022},
+  month = {Oct},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevLett.129.151103},
+  url = {https://link.aps.org/doi/10.1103/PhysRevLett.129.151103}
 }
+
 @article{PhysRevD.107.104036,
 	title        = {Pervasiveness of the breakdown of self-interacting vector field theories},
 	author       = {Coates, Andrew and Ramazano\ifmmode \breve{g}\else \u{g}\fi{}lu, Fethi M.},
