+++ paper/paper.md
@@ -38,8 +38,6 @@ where each sub-Lagranian $\mathcal{L}_n$ contains terms of different orders in t
 
 GRBoondi provides a unified interface for computing the evolution of any generalized Proca model on an arbitrary background. Given a specific background with known expressions for the metric variables and initial data for the Proca field, GRBoondi numerically computes the time evolution of the Proca field, given user-specified generalized Proca equations of motion. While the simultaneous evolution of the metric and matter fields offers the most comprehensive depiction of their temporal progression, there are instances where the density of the Proca field is negligible in comparison to the background curvature. In such cases, fixed background evolution routines serve as an excellent approximation to the complete mutual evolution. A core feature of GRBoondi is that it simplifies many of the boiler-plate code required to begin a simulation from scratch. The only additions a user needs to input are the initial conditions, the modifications to the equations of motion (EOM), and the background functions. GRBoondi will automatically compute various diagnostic quantities and plot files. Moreover, GRBoondi is incredibly modular and modifications to any of the in-built functions is effortless.
 
-The fixed background approximation provides a huge speedup compared to the full evolution of the Einstein equations, since only a small fraction of the variables actually require a time integration. Not only do simulations now require far less resources, the speed up can be up to hundreds of times faster. GRBoondi is based on the publicly available numerical relativity (NR) code GRChombo [@Andrade2021;@Clough2015sqa], which itself is based on the open source adaptive mesh refinement (AMR)-based differential equation solver Chombo [@Adams:2015kgr]. Some smaller pieces of GRBoondi are also based off the recently released open source software GRDzhadzha [@Aurrekoetxea2024], which is also based off of GRChombo. These comprise the core dependencies of GRBoondi.
-
 # Statement of need
 
 In practice, any NR software library can be used to evolve generalized Proca theories. Prominent examples of NR codes encompass the extensive Einstein toolkit [@EinsteinToolkit:2023_11] and its related Cactus framework [@Goodale:2002a], Kranc [@Kranc:web], LEAN [@Sperhake:2006cy], and Canuda [@Canuda:zenodo]. Other libraries worth mentioning are the non-public BAM [@Bruegmann:2006ulg], AMSS-NCKU [@Galaviz:2010mx], PAMR [@East:2011aa] and HAD [@Neilsen:2007ua]. The non-exhaustive list can be expanded with SPeC [@Pfeiffer:2002wt], which is a pseudo-spectral code that uses generalized harmonic coordinates. Similarly, SpECTRE [@deppe_nils_2021_4734670;@Kidder:2016hev;@Cao:2018vhw] uses the Galerkin methods. NRPy [@Ruchlin:2017com] is a Python library that aims for use on non-high performance computing clusters. There are also cosmological simulation codes like CosmoGRaPH [@Mertens:2015ttp] and GRAMSES [@Barrera-Hinojosa:2019mzo]. Simflowny [@Palenzuela:2018sly] is a magneto-hydrodynamic simulation software. GRAthena++ [@Daszuta:2021ecf] uses the oct-tree AMR for maximum scaling. ExaGRyPE is a numerical relativity program leveraging the ExaHyPRE PDE solver [@zhang2024exagrypenumericalgeneralrelativity]. GRDzhadzha [@Aurrekoetxea2024] is another fixed background code, based off the GRChombo [@Andrade2021] framework, which GRBoondi utilizes.
@@ -62,27 +60,15 @@ Since GRBoondi is very similar to the GRChombo code, simulations performed with 
 
[…]
