### 2024-10-30 #6888 @ShaunFell
Hi @warrickball , thank you for your comments. Is 1360 acceptable? I kind of running bare bones on some of the explanations.

> On a more technical note, the citation to the arXiv version of Coates & Ramazanoğlu (2022) can be replaced with the [published version](https://ui.adsabs.harvard.edu/abs/2022PhRvL.129o1103C).

Updated! Thanks for catching that
