### 2024-10-31 9442b40
Initial set of editorial tweaks for JOSS publication
