### 2024-10-31 #6888 @warrickball
Hi @ShaunFell — Thanks for getting the word count down. I've created an [initial PR](https://github.com/ShaunFell/GRBoondi/pull/22) with a few small changes but the references generally need quite a lot of work. In particular, many entries have incomplete author lists, where long lists are replaced with "& others" in the BibTeX. Please at least correct these to contain complete author lists, which should be available either from the relevant journal/archive website, or otherwise [NASA ADS](https://ui.adsabs.harvard.edu/) is a good source for bibliographic information.

If you haven't already, for software you should see if you can find a recommended citation through the relevant documentation but note that this can give mixed quality. For VisIt, for example, there's a [CITATION.cff](https://raw.githubusercontent.com/visit-dav/visit/develop/CITATION.cff) file but the commented BibTeX is actually much richer than the uncommented contents of the file.
