### 2024-10-31 61d72da
update paper.bib according to JOSS review
