### 2024-10-31 #6888 @xuanxu
@ShaunFell the bot command has to be the first line of a comment for the bot to recognize it
