### 2024-11-07 249e22f
Additional editorial tweaks for JOSS publication
