+++ paper/paper.bib
@@ -14,7 +14,7 @@
 }
 
 @article{Heisenberg_2014,
-	title        = {Generalization of the Proca Action},
+	title        = {Generalization of the {Proca} Action},
 	author       = {Heisenberg, Lavinia},
 	year         = 2014,
 	month        = may,
@@ -71,7 +71,7 @@
 }
 
 @article{Clough_2022,
-	title        = {Ghost Instabilities in Self-Interacting Vector Fields: The Problem with Proca Fields},
+	title        = {Ghost Instabilities in Self-Interacting Vector Fields: The Problem with {Proca} Fields},
 	author       = {Clough, Katy and Helfer, Thomas and Witek, Helvi and Berti, Emanuele},
 	year         = 2022,
 	month        = oct,
@@ -116,7 +116,7 @@
 }
 
 @article{_nl_t_rk_2023,
-	title        = {Loss of hyperbolicity and tachyons in generalized Proca theories},
+	title        = {Loss of hyperbolicity and tachyons in generalized {Proca} theories},
 	author       = {\"Unl\"ut\"urk, K{\i}van\c{c} \.I. and Coates, Andrew and Ramazano\u{g}lu, Fethi M.},
 	year         = 2023,
 	month        = aug,
@@ -130,13 +130,13 @@
 }
 
 @article{GENG2021100819,
-	title        = {A dark energy model from generalized Proca theory},
+	title        = {A dark energy model from generalized {Proca} theory},
 	author       = {Chao-Qiang Geng and Yan-Ting Hsu and Jhih-Rong Lu and Lu Yin},
 	year         = 2021,
 	journal      = {Physics of the Dark Universe},
 	volume       = 32,
 	pages        = 100819,
-	doi          = {https://doi.org/10.1016/j.dark.2021.100819},
+	doi          = {10.1016/j.dark.2021.100819},
 	issn         = {2212-6864},
 	url          = {https://www.sciencedirect.com/science/article/pii/S2212686421000509},
 	abstract     = {We consider a specific dark energy model, which only includes the Lagrangian up to the cubic order in terms of the vector field self-interactions in the generalized Proca theory. We examine the cosmological parameters in the model by using the data sets of CMB and CMB+HST, respectively. In particular, the Hubble constant is found to be H0=71.80-0.72+1.07 (72.48-0.60+0.72) kms-1Mpc-1 at 68\%~C.L. with CMB (CMB+HST), which would alleviate the Hubble constant tension. We also obtain that the reduced \ensuremath{\chi}2 values in our model are close to unity when fitting with CMB and CMB+HST, illustrating that our model is a good candidate to describe the cosmological evolutions of the universe.}
@@ -243,7 +243,7 @@
 @Misc{EinsteinToolkit:2023_11,
   requested-for ={EinsteinToolkit},
[…]
