### 2024-11-07 62c94b5
define NR at first occurrence of numerical relativity
