### 2024-11-07 #6888 @warrickball
Hi @ShaunFell — I've just returned to this and opened [another PR](https://github.com/ShaunFell/GRBoondi/pull/23) with what I hope are some final changes. Once those are merged I'll give the paper publication worfklow another test run.
