### 2024-11-07 f218338
Additional editorial tweaks for JOSS publication (#23)
* Additional editorial tweaks for JOSS publication

* define NR at first occurrence of numerical relativity
