### 2024-11-07 #6888 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
